pyproject.toml
[project]
name = "finflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Etienne",email = "etienne.micheladj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "asyncio (>=3.4.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "supabase (>=2.18.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
